@@ -3,7 +3,7 @@ import ApolloAPI
 #endif
 
 @propertyWrapper
-public struct Field<T> {
+public struct Field<T>: Sendable {
 
   let key: StaticString
 
